UI Script Work - "Sailing with the Gods" Project
In this script, I added to existing code to create a new panel for users/players to go in and change their resolution of the game. The script goes through the top five resolution sizes that the current device can support, and presents the top five different resolutions with their highest possible frames-per-second run times. (I got quite a bit of help with coding how the resolutions are chosen.) I also added in a small convenience features for users: the ability to track and see what resolution the game is currently being run in. This is done so by having the current resolution's title in the panel highlight in green font, rather than black, as all other possible options would be.