MiniGameManager - "Sailing with the Gods" Project
For this portion of this project, I created the base code for this current script. I also added in key features for this script to work properly within the game, itself. 

This script works as a controller for specifically the card-battle-type minigame within the SWTG projects that I helped work on. Key features I worked on in this script include: checking the current zone of the map the player is in to get the corresponding type of pirate that goes with that zone; setting up the difficulty of the pirates, which affects negotiations and fleeing chances; having the "Fight" function work properly (i.e. having the player's crew members fight the pirate straight across from them, if there are any);  creating a randomized list of items/materials/drachma (coin) the pirates want if a player chooses to negotiate.