UI and Gameplay Designs - "Sailing with the Gods" Project
Here, I have added in four screen shots of the designs I set up for this project. 

For the Resolutions Settings Screen, I focused on making sure the addition of resolution options fit in with the rest of the menu's appearance and feeling. I mainly wanted it to feel like any other screen that popped up (such as the credits screen). 

The Zones screen shot ties into the Pirate Minigame that my team worked on, specifically the MiniGameManager script, which can be found on the previous page. I referenced a map that was provided to me by the clients, and I created different zones (each one a different color) that represent different areas of the map where different pirates reside. If a player is in a green zone, and only a green zone, they will only run into the pirates that respectively go with the green areas. If a player is in an areas of the map that falls within two or more zones, they will have the chance to run into pirates from any of the zones that they are in. (Not more than one type of pirate in each battle, though, as they are also considered enemies amongst themselves.) Lastly, if a player were to find themselves outside of all the current in-game zones, the game will not run the pirate attack minigame (as there are no pirates in that area to attack the player).  These can also be observed in the MiniGameManager script on the previous page.)

I created the main layout for both of the minigames, as they are shown. Originally, I sketched them out with paper and pencil, and presented that to my team. They liked my ideas, and so I turned paper into product.