Tutorial Design - Capstone Project
For the tutorial level of this project, I completely created the bedroom where the player starts this game. I also added in the wood texture to the floors and stairs and the olive green color to the walls in this scene, as the original layout of this house did not have anything on the floors or walls. I also placed a handful of the objects the player interacts with, and did all of the lighting effects.