Level Design - Capstone Project
The hub world in this project ended up being what was tying all of the individual levels together. To do this, I pulled shaders and textures from each of the levels, and found ways to add them into the floors. I also added in props that can be found in each level near the portal that takes the player to each level. Creating differences in sizing and rotations for these props makes it feel as though there are a lot more variety than there truly is, as when players come closer to objects, they "appear out of thin air" (increase in size until they are at 100% ratio). (I did not write the script for this feature, but I did use it heavily, as can be seen in the screen shots.) The snow particle effect was chosen to also help tie in the feeling of a dream land, and to help tie in the same effect that is used in the main menu scene.