Game Play Level Design - Capstone Project
In this project, I created the main visuals for one of the levels. I was able to take a school prefab that exited on the Unity Asset Store, and rework it to fit better into the game. I added in many trees, a false wall that looks like more trees, and hedges that surround the school to the outside area. For the inside of the school, I moved and deleted some props around in classrooms, placed enemies throughout the school, added in benches/soda and snack machines/plants/lights to the main hallway, and even used some of the props from the school to make new objects, like the double doors that lead to the gym (not pictured).