Cave Level Design - Capstone Project
This is a cave design that I made from different cave pieces. The pieces are individually a floor with walls around it. To create the effect of a cave/underground setting, I took the caves pieces, and flipped them upside down, to make ceilings. I also reworked the shaders for the cave pieces, and created the torches' flame shader by myself. All the lighting in this cave prefab is also mapped out by me.